requirements.txt
alembic==1.16.1
attrs==25.3.0
backports-datetime-fromisoformat==2.0.3
blinker==1.9.0
click==8.2.1
coverage==7.8.2
exceptiongroup==1.3.0
flasgger==0.9.7.1
Flask==3.1.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==4.0.0
mistune==3.1.3
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytest==8.3.5
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
six==1.17.0
SQLAlchemy==2.0.41
tomli==2.2.1
typing_extensions==4.13.2
Werkzeug==3.1.3
